pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["Ivan Krasnoshchok <ivankras67@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
python-multipart = "^0.0.6"
pika = "^1.3.1"
boto3 = "^1.26.142"
python-magic = "^0.4.27"
pandas = "^2.0.1"
openpyxl = "^3.1.2"
PyJWT = "^2.7.0"
requests = "^2.31.0"
dt = "^1.1.61"
attrs = "^23.1.0"
cryptography = "^40.0.2"
pyarrow = "^12.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
